refactor(tests): extract shared Semgrep parity harness, fix language count (#408)
Extract duplicated command building, output parsing, finding
normalization, and skip helpers from the four language-specific
parity test files into tests/common/semgrep_parity_harness.rs.
Each test file now imports the shared harness via `mod common`.

The inverse parity suite also reuses write_file, normalize_path,
semgrep_bin, and skip_if_semgrep_missing from the shared module
while keeping its own NormalizedFinding type (rule_id-based).

Fix language count drift: README.md (lines 36, 172) and
action/action.yml said "10 languages" while Cargo.toml said 11.
All three now say 11.

Closes #408

none
